requirements.txt
python-telegram-bot==20.7
tweepy==4.14.0
openai==1.12.0
sqlalchemy==2.0.25
apscheduler==3.10.4
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2024.1
aiohttp==3.9.1
